feat: グラフのカーソル移動の際のマーカー表示を統一
