+++ src/pages/RouteEditor/tempMarkerIcon.ts
